+++ src/main/java/szewek/mcflux/api/IEnergyConsumer.java
@@ -5,8 +5,8 @@ public interface IEnergyConsumer extends IEnergyHolder {
 	 * Adds energy to a consumer.
 	 * 
 	 * @param amount Energy amount available to be consumed.
-	 * @param simulate If {@code true}, then consumer doesn't change its energy value.
+	 * @param sim If {@code true}, then consumer doesn't change its energy value.
 	 * @return Amount of energy consumed.
 	 */
-	int consumeEnergy(int amount, boolean simulate);
+	int consumeEnergy(int amount, boolean sim);
 }